requirements.txt
Pillow==8.1.0
aiortc==1.0.0
av==8.0.2
numpy==1.19.5
opencv_python==4.5.1.48
streamlit==0.74.1
streamlit_webrtc